cập nhật product service

+++ backend/category-service/src/main/java/com/backend/category_service/service/CategoryService.java
@@ -9,12 +9,15 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.backend.category_service.client.ProductServiceClient;
 import com.backend.category_service.dto.request.CategoryRequest;
 import com.backend.category_service.dto.response.CategoryResponse;
+import com.backend.category_service.dto.response.ProductResponse;
 import com.backend.category_service.entity.Category;
 import com.backend.category_service.mapper.CategoryMapper;
 import com.backend.category_service.repository.CategoryRepository;
@@ -28,11 +31,14 @@ import jakarta.persistence.EntityNotFoundException;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final ProductServiceClient productServiceClient;
     private final Cloudinary cloudinary;
 
-    public CategoryService(CategoryRepository categoryRepository, Cloudinary cloudinary) {
+    public CategoryService(CategoryRepository categoryRepository, Cloudinary cloudinary,
+            ProductServiceClient productServiceClient) {
         this.categoryRepository = categoryRepository;
         this.cloudinary = cloudinary;
+        this.productServiceClient = productServiceClient;
     }
 
     // lấy tất cả category phân trang
@@ -98,7 +104,7 @@ public class CategoryService {
         Category savedCategory = categoryRepository.save(category);
 
         if (image != null && !image.isEmpty()) {
-            String imageUrl = uploadImage(image, savedCategory.getId());
+            String imageUrl = uploadImageOnCloudinary(image, savedCategory.getId());
             savedCategory.setImage(imageUrl);
         }
 
@@ -125,20 +131,38 @@ public class CategoryService {
         category.setSlug(SlugUtil.toSlug(category.getName()));
 
         if (image != null && !image.isEmpty()) {
-            String imageUrl = uploadImage(image, category.getId());
+            String imageUrl = uploadImageOnCloudinary(image, category.getId());
             category.setImage(imageUrl);
         }
 
         return CategoryMapper.toResponse(category);
     }
 
-    // cập nhật status category
+    // cập nhật status
+    @Transactional
     public CategoryResponse updateCategoryStatus(String id, Integer status) {
 
         Category category = categoryRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Danh mục không tồn tại"));
 
         category.setStatus(status);
+        categoryRepository.save(category);
+
+        // cập nhật status category ẩn thì các sản phẩm thuộc category đó sẽ ẩn theo
+        if (status == 0) {
+            ResponseEntity<List<ProductResponse>> response = productServiceClient.getAllActiveProductsByCategoryId(id);
+
+            List<ProductResponse> products = response.getBody();
+
+            if (products != null && !products.isEmpty()) {
+                for (ProductResponse product : products) {
+                    productServiceClient.updateProductStatus(
+                            product.getId(),
+                            0);
+                }
+            }
+        }
+
         return CategoryMapper.toResponse(category);
     }
 
@@ -149,12 +173,19 @@ public class CategoryService {
         Category category = categoryRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Danh mục không tồn tại"));
 
-        deleteFolder(category.getId());
+        Boolean isUsed = productServiceClient.existsByCategoryId(id);
+
+        if (Boolean.TRUE.equals(isUsed)) {
+            throw new IllegalStateException(
+                    "Danh mục này không thể xóa vì đang được sử dụng bởi sản phẩm");
+        }
+
+        deleteFolderOnCloudinary(category.getId());
 
         categoryRepository.delete(category);
     }
 
-    private void deleteFolder(String categoryId) {
+    private void deleteFolderOnCloudinary(String categoryId) {
         try {
             String folderPath = "nckh/categories/" + categoryId;
 
@@ -171,7 +202,7 @@ public class CategoryService {
         }
     }
 
-    private String uploadImage(MultipartFile file, String categoryId) {
+    private String uploadImageOnCloudinary(MultipartFile file, String categoryId) {
 
         // validate size (2MB)
         if (file.getSize() > 2 * 1024 * 1024) {
